requirements.txt
aioschedule==1.0.0
requests==2.26.0
motor==3.6.0
python-dotenv==0.19.0
